Power of an integer

+++ 0x08-recursion/4-pow_recursion.c
@@ -11,7 +11,7 @@
 
 int _pow_recursion(int x, int y)
 {
-	int res = pow(x, y);
+	int res = (int)pow(x, y);
 
 	if (y < 0)
 		return (-1);
